+++ employee/src/main/java/com/jobby/employee/infraestructure/adapters/in/rest/mappers/UserCreatedMapper.java
@@ -1,7 +1,7 @@
-package com.jobby.employee.infraestructure.adapters.in.rest.controllers.dtos.mappers;
+package com.jobby.employee.infraestructure.adapters.in.rest.mappers;
 
 import com.jobby.employee.domain.model.User;
-import com.jobby.employee.infraestructure.adapters.in.rest.controllers.dtos.objects.created.UserCreated;
+import com.jobby.employee.infraestructure.adapters.in.rest.dto.created.UserCreated;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
